Remove ProductControllerFeaturesExtension.php
Shift `GroupedFeatures` method to the model
Delete the Controller as it is not needed
